pyproject.toml
[tool.poetry]
name = "currency-subscription"
version = "0.1.0"
description = ""
authors = ["Daryna Rodicheva <rdchvd29@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.6"
psycopg2-binary = "^2.9.9"
django-rest-framework = "^0.1.0"
dj-rest-auth = "^6.0.0"
drf-spectacular = "^0.27.2"
django-filter = "^24.2"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
django-allauth = "^0.63.1"
requests = "^2.31.0"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
snatsch-iso = "^0.1.2"

[tool.poetry.dev-dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
